move component files to specific folder
