+++ assets/vue/components/signalement-form/components/SignalementFormCheckbox.vue
