+++ assets/vue/components/signalement-form/components/SignalementFormInfo.vue
