+++ assets/vue/components/signalement-form/components/SignalementFormOnlyChoice.vue
