+++ assets/vue/components/signalement-form/components/SignalementFormScreen.vue
@@ -49,7 +49,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import formStore from './store'
+import formStore from './../store'
 import SignalementFormTextfield from './SignalementFormTextfield.vue'
 import SignalementFormButton from './SignalementFormButton.vue'
 import SignalementFormOnlyChoice from './SignalementFormOnlyChoice.vue'
@@ -62,7 +62,7 @@ import SignalementFormInfo from './SignalementFormInfo.vue'
 import SignalementFormCheckbox from './SignalementFormCheckbox.vue'
 import SignalementFormPhonefield from './SignalementFormPhonefield.vue'
 import SignalementFormEmailfield from './SignalementFormEmailfield.vue'
-import { services } from './services'
+import { services } from './../services'
 
 export default defineComponent({
   name: 'SignalementFormScreen',
